refactor: move toc from pkg to internal
